📁 chore(frontend): reorganize JS structure into modules under shared/utils
