+++ src/main/resources/static/js/modules/auth/login.js
@@ -10,7 +10,10 @@
  * @author Jason
  */
 
-import { showToast } from '../../shared/utils/ui/index.js'
+import {
+	showToast,
+	togglePasswordVisibility,
+} from '../../shared/utils/ui/index.js'
 
 $(document).ready(function () {
 	let formSubmitted = false
@@ -113,4 +116,6 @@ $(document).ready(function () {
 			$('#loginText').removeClass('d-none')
 		}
 	})
+
+	togglePasswordVisibility()
 })
