+++ src/main/resources/static/js/modules/authors/authors.js
@@ -17,9 +17,13 @@
 import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
+	isValidText,
+	isValidBirthDate,
+	isValidImageFile,
+	validateImageFileUI,
 	loadSelectOptions,
 	populateSelect,
-} from '../../shared/utils/forms/select-options.js'
+} from '../../shared/utils/forms/index.js'
 
 import {
 	showToast,
@@ -33,13 +37,6 @@ import {
 	initializeTooltips,
 	getCurrentPeruDate,
 } from '../../shared/utils/ui/index.js'
-
-import {
-	isValidText,
-	isValidBirthDate,
-	isValidImageFile,
-	validateImageFileUI,
-} from '../../shared/utils/validators/index.js'
 
 /*****************************************
  * GLOBAL VARIABLES AND HELPER FUNCTIONS
