+++ src/main/resources/static/js/modules/courses/courses.js
@@ -15,6 +15,7 @@
  */
 
 import { loadTableData } from '../../shared/utils/tables/index.js'
+import { isValidText } from '../../shared/utils/forms/index.js'
 
 import {
 	showToast,
@@ -25,8 +26,6 @@ import {
 	setupBootstrapSelectDropdownStyles,
 	initializeTooltips,
 } from '../../shared/utils/ui/index.js'
-
-import { isValidText } from '../../shared/utils/validators/index.js'
 
 /*****************************************
  * TABLE HANDLING
