+++ src/main/resources/static/js/modules/loans/loans.js
@@ -17,9 +17,11 @@
 import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
+	isValidReturnDate,
+	isValidLoanQuantity,
 	loadSelectOptions,
 	populateSelect,
-} from '../../shared/utils/forms/select-options.js'
+} from '../../shared/utils/forms/index.js'
 
 import {
 	showToast,
@@ -32,11 +34,6 @@ import {
 	initializeTooltips,
 	getCurrentPeruDate,
 } from '../../shared/utils/ui/index.js'
-
-import {
-	isValidReturnDate,
-	isValidLoanQuantity,
-} from '../../shared/utils/validators/index.js'
 
 /*****************************************
  * GLOBAL VARIABLES AND HELPER FUNCTIONS
