+++ src/main/resources/static/js/modules/publishers/publishers.js
@@ -17,9 +17,13 @@
 import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
+	isValidText,
+	isValidFoundationYear,
+	isValidImageFile,
+	validateImageFileUI,
 	loadSelectOptions,
 	populateSelect,
-} from '../../shared/utils/forms/select-options.js'
+} from '../../shared/utils/forms/index.js'
 
 import {
 	showToast,
@@ -31,13 +35,6 @@ import {
 	setupBootstrapSelectDropdownStyles,
 	initializeTooltips,
 } from '../../shared/utils/ui/index.js'
-
-import {
-	isValidText,
-	isValidFoundationYear,
-	isValidImageFile,
-	validateImageFileUI,
-} from '../../shared/utils/validators/index.js'
 
 /*****************************************
  * GLOBAL VARIABLES AND HELPER FUNCTIONS
