+++ src/main/resources/static/js/modules/students/students.js
@@ -17,9 +17,15 @@
 import { loadTableData } from '../../shared/utils/tables/index.js'
 
 import {
+	isValidDNI,
+	isValidText,
+	isValidAddress,
+	isValidPhone,
+	isValidEmail,
+	isValidBirthDate,
 	loadSelectOptions,
 	populateSelect,
-} from '../../shared/utils/forms/select-options.js'
+} from '../../shared/utils/forms/index.js'
 
 import {
 	showToast,
@@ -32,15 +38,6 @@ import {
 	initializeTooltips,
 	getCurrentPeruDate,
 } from '../../shared/utils/ui/index.js'
-
-import {
-	isValidDNI,
-	isValidText,
-	isValidAddress,
-	isValidPhone,
-	isValidEmail,
-	isValidBirthDate,
-} from '../../shared/utils/validators/index.js'
 
 /*****************************************
  * GLOBAL VARIABLES AND HELPER FUNCTIONS
