+++ src/main/resources/static/js/shared/utils/ui/index.js
@@ -1,235 +1,8 @@
-export function showToast(message, type) {
-	const $toastContainer = $('#toast-container')
-	$toastContainer.find('.toast').toast('hide').remove()
-
-	let borderClasses = ''
-	let iconClass = ''
-	let btnCloseClass = ''
-
-	switch (type) {
-		case 'success':
-			borderClasses =
-				'text-success-emphasis bg-success-subtle border border-success-subtle rounded-2'
-			iconClass = 'bi-check2-circle'
-			btnCloseClass = 'btn-close-success'
-			break
-		case 'error':
-			borderClasses =
-				'text-danger-emphasis bg-danger-subtle border border-danger-subtle rounded-2'
-			iconClass = 'bi-x-circle'
-			btnCloseClass = 'btn-close-danger'
-			break
-		case 'warning':
-			borderClasses =
-				'text-warning-emphasis bg-warning-subtle border border-warning-subtle rounded-2'
-			iconClass = 'bi-exclamation-triangle'
-			btnCloseClass = 'btn-close-warning'
-			break
-	}
-
-	const $toastElement = $(`
-		<div class="toast align-items-center ${borderClasses}" role="alert" aria-live="assertive" aria-atomic="true">
-			<div class="d-flex">
-				<div class="toast-body d-flex align-items-center">
-					<i class="bi ${iconClass} me-2"></i>${message}
-				</div>
-				<button type="button" class="btn-close ${btnCloseClass} me-2 m-auto" data-bs-dismiss="toast" aria-label="Close"></button>
-			</div>
-		</div>
-	`)
-
-	$toastContainer.append($toastElement)
-
-	const bootstrapToast = new bootstrap.Toast($toastElement[0], {
-		animation: true,
-		autohide: true,
-		delay: 3000,
-	})
-	bootstrapToast.show()
-
-	$toastElement.on('hidden.bs.toast', function () {
-		$(this).remove()
-	})
-}
-
-export function toggleButtonLoading(button, loading = true) {
-	const $button = $(button)
-	const $icon = $button.children('i')
-	const $spinner = $button.children('.spinner-border')
-
-	if (loading) {
-		$button.prop('disabled', true)
-		$icon.addClass('d-none')
-		$spinner.removeClass('d-none')
-	} else {
-		$button.prop('disabled', false)
-		$icon.removeClass('d-none')
-		$spinner.addClass('d-none')
-	}
-}
-
-export function toggleModalLoading(triggerElement, loading = true) {
-	const modalSelector = $(triggerElement).data('bs-target')
-	const $modal = $(modalSelector)
-
-	const $spinner = $modal.find('[id$="Spinner"]')
-	const $content = $modal.find('[id$="Form"], [id$="Content"]')
-	const $button = $modal.find('button[id$="Btn"]')
-
-	if (loading) {
-		$spinner.removeClass('d-none')
-		$content.addClass('d-none')
-		if ($button.length) $button.prop('disabled', true)
-	} else {
-		$spinner.addClass('d-none')
-		$content.removeClass('d-none')
-		if ($button.length) $button.prop('disabled', false)
-	}
-}
-
-export function placeholderColorSelect() {
-	$('select.selectpicker').on('change', function () {
-		const $select = $(this)
-		const $dropdown = $select.closest('.bootstrap-select')
-		const $filterOption = $dropdown.find('.filter-option-inner-inner')
-
-		if ($select.val() !== '' && $select.val() !== null) {
-			$dropdown.removeClass('placeholder-color')
-			$filterOption.css('color', 'var(--bs-body-color)')
-		}
-	})
-}
-
-export function placeholderColorEditSelect() {
-	$('select[id^="edit"]').each(function () {
-		const $select = $(this)
-		const $dropdown = $select.closest('.bootstrap-select')
-		const $filterOption = $dropdown.find('.filter-option-inner-inner')
-
-		if ($filterOption.text().trim() === 'No hay selección') {
-			$filterOption.css('color', 'var(--placeholder-color)')
-		} else {
-			$filterOption.css('color', 'var(--bs-body-color)')
-		}
-	})
-}
-
-export function placeholderColorDateInput() {
-	$('input[type="date"]').each(function () {
-		const $input = $(this)
-
-		if (!$input.val()) {
-			$input.css('color', 'var(--placeholder-color)')
-		} else {
-			$input.css('color', '')
-		}
-	})
-
-	$('input[type="date"]').on('change input', function () {
-		const $input = $(this)
-
-		if (!$input.val()) {
-			$input.css('color', 'var(--placeholder-color)')
-		} else {
-			$input.css('color', '')
-		}
-	})
-}
-
-export function initializeCropper(
-	file,
-	$cropperContainer,
-	$imageToCrop,
-	cropper,
-) {
-	const reader = new FileReader()
-	reader.onload = function (e) {
-		$cropperContainer.removeClass('d-none')
-		$imageToCrop.attr('src', e.target.result)
-
-		if (cropper) {
-			cropper.destroy()
-		}
-
-		cropper = new Cropper($imageToCrop[0], {
-			aspectRatio: 1,
-			viewMode: 1,
-			autoCropArea: 1,
-			responsive: true,
-			checkOrientation: false,
-			ready: function () {
-				$('.cropper-crop-box').css({
-					'border-radius': '50%',
-					overflow: 'hidden',
-				})
-			},
-		})
-	}
-	reader.readAsDataURL(file)
-}
-
-function updateDropdownIcons($dropdown) {
-	$dropdown.find('.dropdown-item').each(function () {
-		const $item = $(this)
-		const $icon = $item.find('i.bi-check2')
-
-		if ($item.hasClass('active') && $item.hasClass('selected')) {
-			if ($icon.length === 0) {
-				$('<i class="bi bi-check2 ms-auto"></i>').appendTo($item)
-			}
-		} else {
-			$icon.remove()
-		}
-	})
-}
-
-export function setupBootstrapSelectDropdownStyles() {
-	const observer = new MutationObserver((mutationsList) => {
-		mutationsList.forEach((mutation) => {
-			mutation.addedNodes.forEach((node) => {
-				if (node.nodeType === 1 && node.classList.contains('dropdown-menu')) {
-					const $dropdown = $(node)
-					$dropdown.addClass('gap-1 px-2 rounded-3 mx-0 shadow')
-					$dropdown
-						.find('.dropdown-item')
-						.addClass(
-							'rounded-2 d-flex align-items-center justify-content-between',
-						)
-
-					$dropdown.find('li:not(:first-child)').addClass('mt-1')
-
-					updateDropdownIcons($dropdown)
-				}
-			})
-		})
-	})
-
-	observer.observe(document.body, { childList: true, subtree: true })
-
-	$(document).on('click', '.bootstrap-select .dropdown-item', function () {
-		const $dropdown = $(this).closest('.dropdown-menu')
-		updateDropdownIcons($dropdown)
-	})
-}
-
-export function initializeTooltips(container) {
-	$(container)
-		.find('[data-tooltip="tooltip"]')
-		.tooltip({
-			trigger: 'hover',
-		})
-		.on('click', function () {
-			$(this).tooltip('hide')
-		})
-}
-
-export function getCurrentPeruDate() {
-	const peruDateStr = new Date().toLocaleString('en-US', {
-		timeZone: 'America/Lima',
-		year: 'numeric',
-		month: '2-digit',
-		day: '2-digit',
-	})
-	const [month, day, year] = peruDateStr.split('/')
-	return new Date(`${year}-${month}-${day}`)
-}
+export * from './toast.js'
+export * from './loading.js'
+export * from './select-style.js'
+export * from './date-input-style.js'
+export * from './cropper.js'
+export * from './tooltip.js'
+export * from './time-utils.js'
+export * from './password.js'
