+++ src/main/resources/static/js/shared/utils/ui/tooltip.js
@@ -1,0 +1,10 @@
+export function initializeTooltips(container) {
+	$(container)
+		.find('[data-tooltip="tooltip"]')
+		.tooltip({
+			trigger: 'hover',
+		})
+		.on('click', function () {
+			$(this).tooltip('hide')
+		})
+}
